add CUD item tests

+++ src/defResolver/item.defResolver.ts
@@ -6,13 +6,16 @@ import { ItemCreate, ItemUpdate } from "../interfaces/itemInterfaces";
 import { StatusEnum, CategoryEnum } from "../interfaces/enums";
 import { Image } from "../entity/Image";
 import { fileSaver } from "../utils/saveFile";
-import { validateItemCreate } from "../validation/itemValidation";
+import {
+  validateItemCreate,
+  validateItemUpdate,
+} from "../validation/itemValidation";
 
 export const getItemQuery = async (_: any, args: { id: number }) => {
   const { id } = args;
   try {
     const item = await Item.findOne(id, {
-      relations: ["images", "giver", "wishers", "taker"]
+      relations: ["images", "giver", "wishers", "taker"],
     });
     if (!item) throw new Error("Item not found");
     return item;
@@ -23,7 +26,7 @@ export const getItemQuery = async (_: any, args: { id: number }) => {
 
 export const getItemsQuery = async (_: any, args: ItemSearch) => {
   const {
-    input: { skip, first, name, userId, view, category = CategoryEnum.ALL }
+    input: { skip, first, name, userId, view, category = CategoryEnum.ALL },
   } = args;
   try {
     const base = Item.createQueryBuilder("item")
@@ -34,37 +37,37 @@ export const getItemsQuery = async (_: any, args: ItemSearch) => {
       .leftJoinAndSelect("item.images", "image");
     if (view === "items")
       base.andWhere("item.status = :status", {
-        status: StatusEnum.ONGOING
+        status: StatusEnum.ONGOING,
       });
     if (name) {
       base.andWhere("LOWER(item.name) like LOWER(:name)", {
-        name: `%${name}%`
+        name: `%${name}%`,
       });
       base.andWhere("item.status = :status", {
-        status: StatusEnum.ONGOING
+        status: StatusEnum.ONGOING,
       });
     }
     if (category !== CategoryEnum.ALL) {
       base.andWhere("item.category = :category", {
-        category: category
+        category: category,
       });
     }
     if (view === "giving") {
       base.andWhere("item.giverId = :giverId", { giverId: userId });
       base.andWhere("item.status != :status", {
-        status: StatusEnum.GIVEN
+        status: StatusEnum.GIVEN,
       });
     }
     if (view === "taken") {
       base.andWhere("item.status = :status", {
-        status: StatusEnum.GIVEN
+        status: StatusEnum.GIVEN,
       });
       base.andWhere("item.takerId = :takerId", { takerId: userId });
     }
     if (view === "given") {
       base.andWhere("item.giverId = :giverId", { giverId: userId });
       base.andWhere("item.status = :status", {
-        status: StatusEnum.GIVEN
+        status: StatusEnum.GIVEN,
       });
     }
     const items = await base.getMany();
@@ -76,7 +79,7 @@ export const getItemsQuery = async (_: any, args: ItemSearch) => {
 
 export const countItemsQuery = async (_: any, args: ItemCount) => {
   const {
-    input: { takerId, view, category = CategoryEnum.ALL }
+    input: { takerId, view, category = CategoryEnum.ALL },
   } = args;
   try {
     const base = Item.createQueryBuilder("item").select("COUNT(item)", "count");
@@ -93,7 +96,7 @@ export const countItemsQuery = async (_: any, args: ItemCount) => {
     else if (view === "giving")
       base
         .where("item.status != :status", {
-          status: StatusEnum.GIVEN
+          status: StatusEnum.GIVEN,
         })
         .andWhere("item.giverId = :giverId", { giverId: takerId });
     if (category !== CategoryEnum.ALL)
@@ -105,15 +108,60 @@ export const countItemsQuery = async (_: any, args: ItemCount) => {
   }
 };
 
+export const addItemsMutation = async (
+  _: any,
+  args: { item: ItemCreate; file: any },
+  context: any
+) => {
+  try {
+    const { req } = context;
+    if (!req.isAuth) throw new Error("Unauthorized!");
+    const { item, file } = args;
+    validateItemCreate(item);
+    const { name, active, status, state, category, description } = item;
+    const itemNew = Item.create({
+      name,
+      active,
+      status,
+      state,
+      category,
+      description,
+      giverId: req.userId,
+    });
+    if (file) {
+      await getConnection().transaction(async (transactionalEntityManager) => {
+        const imageUrl = await fileSaver(file);
+        const savedItem = await transactionalEntityManager.save(item);
+        const newImage = Image.create({
+          url: imageUrl,
+          item: savedItem,
+        });
+        await transactionalEntityManager.save(newImage);
+      });
+    } else {
+      await itemNew.save();
+    }
+    return itemNew;
+  } catch (err) {
+    throw new Error(err ? err : "Server error!");
+  }
+};
+
 export const updateItemQuery = async (
   _: any,
-  args: { id: number; item: ItemUpdate }
+  args: { id: number; item: ItemUpdate },
+  context: any
 ) => {
-  const { id, item } = args;
-  const { name, active, status, state, category, description } = item;
   try {
+    const { req } = context;
+    if (!req.isAuth) throw new Error("Unauthorized!");
+    const { id, item } = args;
+    validateItemUpdate(item);
+    const { name, active, status, state, category, description } = item;
     const updateItem = await Item.findOne(id);
     if (!updateItem) throw new Error("Item does not exist!");
+    if (req.userId !== updateItem.giverId)
+      throw new Error("Cannot change other user item!");
     if (updateItem.status === StatusEnum.GIVEN)
       throw new Error("Item alread given!");
     if (name) updateItem.name = name;
@@ -129,55 +177,24 @@ export const updateItemQuery = async (
   }
 };
 
-export const deleteItemQuery = async (_: any, args: { id: number }) => {
-  const { id } = args;
+export const deleteItemQuery = async (
+  _: any,
+  args: { id: number },
+  context: any
+) => {
   try {
+    const { req } = context;
+    if (!req.isAuth) throw new Error("Unauthorized!");
+    const { id } = args;
     const deleteItem = await Item.findOne(id);
     if (!deleteItem) throw new Error("Item does not exist!");
+    if (req.userId !== deleteItem.giverId)
+      throw new Error("Cannot delete other user item!");
     if (deleteItem.status === "given")
       throw new Error("Cannot delete archived item!");
     const returnItem = { ...deleteItem };
     await deleteItem.remove();
     return returnItem;
-  } catch (err) {
-    throw new Error(err ? err : "Server error!");
-  }
-};
-
-export const addItemsMutation = async (
-  _: any,
-  args: { item: ItemCreate; file: any },
-  context: any
-) => {
-  if (!context.req.isAuth) throw new Error("Unauthorized! You need to log in!");
-  const { item, file } = args;
-  validateItemCreate(item);
-  const { name, active, status, state, category, description, giverId } = item;
-  try {
-    if (!name || !giverId) throw new Error("Provide missing data!");
-    const item = Item.create({
-      name,
-      active,
-      status,
-      state,
-      category,
-      description,
-      giverId
-    });
-    if (file) {
-      await getConnection().transaction(async transactionalEntityManager => {
-        const imageUrl = await fileSaver(file);
-        const savedItem = await transactionalEntityManager.save(item);
-        const newImage = Image.create({
-          url: imageUrl,
-          item: savedItem
-        });
-        await transactionalEntityManager.save(newImage);
-      });
-    } else {
-      await item.save();
-    }
-    return item;
   } catch (err) {
     throw new Error(err ? err : "Server error!");
   }
